Add put product

+++ src/main/java/com/codewithmosh/store/dtos/ProductDto.java
@@ -5,13 +5,12 @@ import lombok.Data;
 import java.math.BigDecimal;
 
 @Data
-
 public class ProductDto {
 
   private Long id;
   private String name;
   private BigDecimal price;
   private String description;
-  private Byte categoryId;
+  private Long categoryId;
 
 }
